requirements.txt
PyGithub
requests
pendulum
openai
pyTelegramBotAPI